[tested] move manifest package into core
