+++ pkg/core/manifest/manifest.go
