+++ pkg/core/manifest/manifest_types.go
