+++ pkg/core/manifest/node.go
