+++ pkg/core/manifest/tests/manifests/colliding_dir_frontmatters.yaml
