+++ pkg/core/manifest/tests/manifests/fileTree_filtering.yaml
