+++ pkg/core/manifest/tests/manifests/index_md_with_properties.yaml
