+++ pkg/core/manifest/tests/manifests/manifest.yaml
