+++ pkg/core/manifest/tests/manifests/merging.yaml
