+++ pkg/core/manifest/tests/results/index_md_with_properties.yaml
