+++ pkg/core/manifest/tests/results/manifest.yaml
