+++ pkg/core/manifest/tests/results/merging.yaml
