+++ pkg/core/run.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 
 	"github.com/gardener/docforge/pkg/core/linkvalidator"
-	"github.com/gardener/docforge/pkg/manifest"
+	"github.com/gardener/docforge/pkg/core/manifest"
 	"github.com/gardener/docforge/pkg/plugins"
 	"github.com/gardener/docforge/pkg/registry"
 	"k8s.io/klog/v2"
