+++ pkg/plugins/docsy.go
@@ -9,7 +9,7 @@ import (
 	"path"
 	"strings"
 
-	"github.com/gardener/docforge/pkg/manifest"
+	"github.com/gardener/docforge/pkg/core/manifest"
 	"github.com/gardener/docforge/pkg/registry"
 )
