+++ pkg/plugins/downloader/plugin.go
@@ -8,7 +8,7 @@ import (
 	"context"
 	"path"
 
-	"github.com/gardener/docforge/pkg/manifest"
+	"github.com/gardener/docforge/pkg/core/manifest"
 	"github.com/gardener/docforge/pkg/plugins"
 	"github.com/gardener/docforge/pkg/registry"
 	"github.com/gardener/docforge/pkg/writers"
