+++ pkg/plugins/markdown/document_worker.go
@@ -13,7 +13,7 @@ import (
 	"sync"
 
 	"github.com/gardener/docforge/cmd/hugo"
-	"github.com/gardener/docforge/pkg/manifest"
+	"github.com/gardener/docforge/pkg/core/manifest"
 	"github.com/gardener/docforge/pkg/plugins/markdown/frontmatter"
 	"github.com/gardener/docforge/pkg/plugins/markdown/linkresolver"
 	"github.com/gardener/docforge/pkg/plugins/markdown/parser"
