+++ pkg/plugins/markdown/linkresolver/link_resolving.go
@@ -12,8 +12,8 @@ import (
 	"strings"
 
 	"github.com/gardener/docforge/cmd/hugo"
+	"github.com/gardener/docforge/pkg/core/manifest"
 	"github.com/gardener/docforge/pkg/internal/link"
-	"github.com/gardener/docforge/pkg/manifest"
 	"github.com/gardener/docforge/pkg/registry"
 	"github.com/gardener/docforge/pkg/registry/repositoryhost"
 	"k8s.io/klog/v2"
