+++ pkg/writers/writersfakes/fake_writer.go
@@ -7,7 +7,7 @@ package writersfakes
 import (
 	"sync"
 
-	"github.com/gardener/docforge/pkg/manifest"
+	"github.com/gardener/docforge/pkg/core/manifest"
 	"github.com/gardener/docforge/pkg/writers"
 )
